noapp.kill file in app dir prevents boot of app
Added logic to check for the presence of /app/noapp.kill during boot. If the file exists, the script logs a message, deletes the file, and skips starting /app/app, similar to the existing gadget mode check.  This includes a self- rm -f to not brick if users somehow lose ssh access, use this in unintended ways, etc.